+++ src/components/TeacherComponents/SubjectListingCard/SubjectListingCard.test.tsx
@@ -3,11 +3,10 @@ import userEvent from "@testing-library/user-event";
 
 import SubjectListingCard from "./SubjectListingCard";
 
-import { Subjects } from "@/pages/teachers/key-stages/[keyStageSlug]/subjects";
 import renderWithTheme from "@/__tests__/__helpers__/renderWithTheme";
 import subjectPagePropsFixture from "@/node-lib/curriculum-api-2023/fixtures/subjectListing.fixture";
 
-const subjects: Subjects = subjectPagePropsFixture().subjects;
+const subjects = subjectPagePropsFixture().subjects;
 
 const subjectSelected = jest.fn();
 jest.mock("@/context/Analytics/useAnalytics", () => ({
@@ -27,7 +26,7 @@ describe("SubjectListingCardDouble", () => {
   test("render a Card with the Name of the Subject", () => {
     renderWithTheme(
       <SubjectListingCard
-        subject={subjects[0] as Subjects[number]}
+        subject={subjects[0]}
         subjectSlug={"biology"}
         keyStageSlug={"ks4"}
         keyStageTitle={"Key stage 4"}
@@ -38,7 +37,7 @@ describe("SubjectListingCardDouble", () => {
   test("new units with 1 programme take you to 'teachers' view unit listing page", () => {
     const { getByRole } = renderWithTheme(
       <SubjectListingCard
-        subject={subjects[0] as Subjects[number]}
+        subject={subjects[0]}
         subjectSlug={"biology"}
         keyStageSlug={"ks4"}
         keyStageTitle={"Key stage 4"}
@@ -54,34 +53,40 @@ describe("SubjectListingCardDouble", () => {
     );
   });
   test("new units are labeled as 'new'", () => {
-    const { getByText } = renderWithTheme(
-      <SubjectListingCard
-        subject={subjects[2] as Subjects[number]}
-        subjectSlug={"maths"}
-        keyStageSlug={"ks4"}
-        keyStageTitle={"Key stage 4"}
-      />,
-    );
-    const cardClickTarget = getByText("New");
-    expect(cardClickTarget).toBeInTheDocument();
+    const math = subjects[2];
+    if (math) {
+      const { getByText } = renderWithTheme(
+        <SubjectListingCard
+          subject={math}
+          subjectSlug={"maths"}
+          keyStageSlug={"ks4"}
+          keyStageTitle={"Key stage 4"}
+        />,
+      );
+      const cardClickTarget = getByText("New");
+      expect(cardClickTarget).toBeInTheDocument();
+    }
   });
   test("new label is not visible on old units", () => {
-    const { queryByText } = renderWithTheme(
-      <SubjectListingCard
-        subject={subjects[3] as Subjects[number]}
-        subjectSlug={"maths"}
-        keyStageSlug={"ks4"}
-        keyStageTitle={"Key stage 4"}
-      />,
-    );
-    const cardClickTarget = queryByText("New");
-    expect(cardClickTarget).not.toBeInTheDocument();
+    const math = subjects[3];
+    if (math) {
+      const { queryByText } = renderWithTheme(
+        <SubjectListingCard
+          subject={math}
+          subjectSlug={"maths"}
+          keyStageSlug={"ks4"}
+          keyStageTitle={"Key stage 4"}
+        />,
+      );
+      const cardClickTarget = queryByText("New");
+      expect(cardClickTarget).not.toBeInTheDocument();
+    }
   });
 
   test("calls tracking.subjectSelected once, with correct props", async () => {
     const { getByRole } = renderWithTheme(
       <SubjectListingCard
-        subject={subjects[0] as Subjects[number]}
+        subject={subjects[0]}
         subjectSlug={"maths"}
         keyStageSlug={"ks4"}
         keyStageTitle={"Key stage 4"}