+++ src/node-lib/curriculum-api-2023/queries/subjectListing/constructSubjectsFromUnitData.ts
@@ -12,6 +12,8 @@ interface UnprocessedSubject {
     unitIds: Set<number>;
     lessonIds: Set<number>;
     programmeSlugs: Set<string>;
+    pathwaySlug: "core" | "gcse" | null;
+    pathwayTitle: "Core" | "GCSE" | null;
   };
 }
 
@@ -22,6 +24,8 @@ interface ProcessedSubject {
   unitCount: number;
   lessonCount: number;
   programmeCount: number;
+  pathwaySlug: "core" | "gcse" | null;
+  pathwayTitle: "Core" | "GCSE" | null;
 }
 
 export const constructSubjectsFromUnitData = (
@@ -38,6 +42,8 @@ export const constructSubjectsFromUnitData = (
       subject_slug,
       tier_slug,
       examboard_slug,
+      pathway_slug,
+      pathway,
     } = unit.programme_fields;
     const { unit_id } = unit_data;
     const originalProgrammeSlug = programme_slug;
@@ -65,6 +71,8 @@ export const constructSubjectsFromUnitData = (
         unitIds: new Set([unit_id]),
         lessonIds: new Set(lesson_ids),
         programmeSlugs: new Set([originalProgrammeSlug]),
+        pathwaySlug: pathway_slug,
+        pathwayTitle: pathway,
       };
     }
 
@@ -80,6 +88,8 @@ export const constructSubjectsFromUnitData = (
       unitCount: subject.unitIds.size,
       lessonCount: subject.lessonIds.size,
       programmeCount: subject.programmeSlugs.size,
+      pathwaySlug: subject.pathwaySlug,
+      pathwayTitle: subject.pathwayTitle,
     }),
   );